Beginning.
To start I have uploaded all of the necessary files to begin with from basic Python, HTML and CSS knowledge to build a web application. I have also added a requirements.txt file as it contains libraries for a data base that will hold information in a task manager. The task manager will be added at a future date. for now these will be the files I will start building.