Update styles.css
In this commit I will be showcasing my current skill within web development by making a relative theme that blend with the portfolio background image. In this example I have chosen an image of a wheat field and moulded the theme colours towards orange and brown, I did originally experiment with the following colours; maroon, gold and beige.